move extractRequestData to @sentry/utils
